Refactor sidebars into base component
