+++ src/AdminSidebar.test.jsx
@@ -42,3 +42,4 @@ test('renders Site Settings tab', () => {
   );
   expect(screen.getByText('Site Settings')).toBeInTheDocument();
 });
+
