+++ src/Sidebar.test.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { render } from '@testing-library/react';
 import { MemoryRouter } from 'react-router-dom';
 import '@testing-library/jest-dom';
-import Sidebar from './Sidebar';
+import SidebarBase from './components/SidebarBase';
 
 jest.mock('./firebase/config', () => ({ auth: {} }));
 jest.mock('firebase/auth', () => ({ signOut: jest.fn() }));
@@ -10,10 +10,11 @@ jest.mock('firebase/auth', () => ({ signOut: jest.fn() }));
 test('sidebar has md width class', () => {
   const { container } = render(
     <MemoryRouter>
-      <Sidebar />
+      <SidebarBase tabs={[{ label: 'Home', path: '/' }]} />
     </MemoryRouter>
   );
   const sidebarDiv = container.querySelector('.border-r');
   expect(sidebarDiv).toHaveClass('w-[250px]');
   expect(sidebarDiv).toHaveClass('md:flex');
 });
+
